+++ src/components/app/app.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+import MainScreen from "../main-screen/main-screen.jsx";
+
+const App = () => {
+  return <MainScreen
+  />;
+};
+
+export default App;
