+++ src/index.js
@@ -1,10 +1,11 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./components/app.jsx";
+import App from "./components/app/app.jsx";
 
 const init = () => {
   ReactDOM.render(
-      <App />,
+      <App
+      />,
       document.querySelector(`#root`)
   );
 };
